chore: place all enums into a single file
