+++ frontend/components/receipts/vote-submission/vote-submission.tsx
@@ -2,16 +2,11 @@ import { vegaVoteValue } from '@vegaprotocol/rest-clients/dist/trading-data'
 import { ReactNode } from 'react'
 
 import { DataTable } from '@/components/data-table/data-table'
+import { VOTE_VALUE_MAP } from '@/components/enums'
 
 import { ProposalLink } from '../../vega-entities/proposal-link'
 import { ReceiptComponentProperties } from '../receipts'
 import { ReceiptWrapper } from '../utils/receipt-wrapper'
-
-const VOTE_VALUE_MAP = {
-  [vegaVoteValue.VALUE_YES]: 'For',
-  [vegaVoteValue.VALUE_NO]: 'Against',
-  [vegaVoteValue.VALUE_UNSPECIFIED]: 'Unspecified'
-}
 
 export const VoteSubmission = ({ transaction }: ReceiptComponentProperties) => {
   const { proposalId, value } = transaction.voteSubmission
