+++ frontend/routes/auth/wallets/key-details/assets-list/asset-card.tsx
@@ -1,44 +1,14 @@
-import { vegaAccount, vegaAccountType } from '@vegaprotocol/rest-clients/dist/trading-data'
+import { vegaAccount } from '@vegaprotocol/rest-clients/dist/trading-data'
 import { addDecimalsFormatNumber, formatNumber, toBigNum } from '@vegaprotocol/utils'
 import BigNumber from 'bignumber.js'
 import get from 'lodash/get'
 
 import { CollapsibleCard } from '@/components/collapsible-card'
 import { DataTable } from '@/components/data-table/data-table'
+import { ACCOUNT_TYPE_MAP } from '@/components/enums'
 import { useAssetsStore } from '@/stores/assets-store'
 
 import { MarketLozenges } from './markets-lozenges'
-
-export const ACCOUNT_TYPE_MAP: Record<vegaAccountType, string> = {
-  [vegaAccountType.ACCOUNT_TYPE_INSURANCE]: 'Insurance',
-  [vegaAccountType.ACCOUNT_TYPE_SETTLEMENT]: 'Settlement',
-  [vegaAccountType.ACCOUNT_TYPE_MARGIN]: 'Margin',
-  [vegaAccountType.ACCOUNT_TYPE_GENERAL]: 'General',
-  [vegaAccountType.ACCOUNT_TYPE_FEES_INFRASTRUCTURE]: 'Fees (infra)',
-  [vegaAccountType.ACCOUNT_TYPE_FEES_LIQUIDITY]: 'Fees (liquidity)',
-  [vegaAccountType.ACCOUNT_TYPE_FEES_MAKER]: 'Fees (maker)',
-  [vegaAccountType.ACCOUNT_TYPE_BOND]: 'Bond',
-  [vegaAccountType.ACCOUNT_TYPE_EXTERNAL]: 'External',
-  [vegaAccountType.ACCOUNT_TYPE_GLOBAL_INSURANCE]: 'Global insurance',
-  [vegaAccountType.ACCOUNT_TYPE_GLOBAL_REWARD]: 'Global reward',
-  [vegaAccountType.ACCOUNT_TYPE_PENDING_TRANSFERS]: 'Pending transfers',
-  [vegaAccountType.ACCOUNT_TYPE_REWARD_MAKER_PAID_FEES]: 'Maker paid fees',
-  [vegaAccountType.ACCOUNT_TYPE_REWARD_MAKER_RECEIVED_FEES]: 'Maker received fees',
-  [vegaAccountType.ACCOUNT_TYPE_REWARD_LP_RECEIVED_FEES]: 'LP received fees',
-  [vegaAccountType.ACCOUNT_TYPE_REWARD_MARKET_PROPOSERS]: 'Market proposers',
-  [vegaAccountType.ACCOUNT_TYPE_UNSPECIFIED]: 'Unspecified',
-  [vegaAccountType.ACCOUNT_TYPE_HOLDING]: 'Holding',
-  [vegaAccountType.ACCOUNT_TYPE_LP_LIQUIDITY_FEES]: 'LP liquidity fees',
-  [vegaAccountType.ACCOUNT_TYPE_LIQUIDITY_FEES_BONUS_DISTRIBUTION]: '',
-  [vegaAccountType.ACCOUNT_TYPE_NETWORK_TREASURY]: 'Network treasury',
-  [vegaAccountType.ACCOUNT_TYPE_VESTING_REWARDS]: 'Vesting rewards',
-  [vegaAccountType.ACCOUNT_TYPE_VESTED_REWARDS]: 'Vested rewards',
-  [vegaAccountType.ACCOUNT_TYPE_REWARD_AVERAGE_POSITION]: 'Average position rewards',
-  [vegaAccountType.ACCOUNT_TYPE_REWARD_RELATIVE_RETURN]: 'Relative return rewards',
-  [vegaAccountType.ACCOUNT_TYPE_REWARD_RETURN_VOLATILITY]: 'Volatility rewards',
-  [vegaAccountType.ACCOUNT_TYPE_REWARD_VALIDATOR_RANKING]: 'Validator ranking rewards',
-  [vegaAccountType.ACCOUNT_TYPE_PENDING_FEE_REFERRAL_REWARD]: 'Pending fee referral rewards'
-}
 
 export const locators = {
   assetCard: 'asset-card',
